Fix token bucket to use a double-based counter internally.
This fixes cases in which the periodic increment is less than 1, e.g. when
the refill time is extremely small.

+++ src/tor/model/tokenbucket.h
@@ -23,7 +23,7 @@ public:
 private:
   void Refill ();
 
-  int64_t m_bucket;
+  double m_bucket;
   DataRate m_rate;
   DataRate m_burst;
   Time m_refilltime;
